Separating out the mismatch reporter from the rest of the merge code.
